Basic support for class proxying.
