+++ tools/objc/interface/macros.tpl
