+++ tools/objc/interface/sink.tpl
